Extract CodeOfConductService
